chore: Factor out playerdata conversion+types
